Correct the meaning of a PR link for a rule (under specification)
Existence of a PR link means the rule specification is not merged to master yet,
and might still be undergoing specification.
It should not be hidden if the rule is already covered by an analyzer.

If the rule is still in the PR it means it's specification is not complete,
and might change before the PR is merged.

The original meaning of "Not implemented" stems from the deprecated policy of
merging the rspec to the master before merging the implementation of the rule
in an analyzer (which itself was caused by the technical limitation of rule-api,
since removed).