Define all_required_checks task as the convenient waiting point (#813)
`update_coverage` action used to wait for `validate_asciidoc` to succeed as the task that takes the longest. After #782 is merged as f21f23b9860c10d4effb00b6022917bc28e02e61 it is no longer the case. This change introduces `all_required_checks` task that depends on all checks required for a PR to be merged, thus providing a single point that is known to succeed last and can be safely waited upon.