Update flask.adoc
I couldn't get Sonarcloud to trigger this issue using the provided noncompliant code example. 

I think the code examples as written end up being circular because the local function 'redirect()' will call itself rather than the imported 'redirect()' function of the same name. The fix is to change the local function name to be redirecting(). I changed the API endpoint name as well so that it matched.