Update flask.adoc
Undo my removal of 'Flask()' calls. This will cause a hotspot to be raised but I don't know enough to fix that properly so will have to leave it for now. I'll leave this PR solving the circular call issue and raise a dogfood post on the compliant code causing a hotspot to appear.