Modify S3394(COBOL): Remove compliant samples (#3415)
## Why

https://github.com/SonarSource/rspec/commit/b4812424ee5b9cd210f926252566ac6c3bb291dd#r131831666
This code sample will raise, as the COBOL logic is pretty simple. The
idea is to make this rule a "review" rule instead of a vulnerability. We
do not have the bandwidth to convert it to a hotspot but that's ok.

## Review

A dedicated reviewer checked the rule description successfully for:

- [ ] logical errors and incorrect information
- [ ] information gaps and missing content
- [ ] text style and tone
- [ ] PR summary and labels follow [the
guidelines](https://github.com/SonarSource/rspec/#to-modify-an-existing-rule)