Update description.adoc to allow for freeform text in 'how' tab
We allow for freeform text in the 'why' tab. We want to allow for this in the 'how' tab.

I have updated the language a bit to talk about 'story-telling' rather than 'blog style' because I think that better reflects how we expect this approach to be used. 

I have also removed the (optional) markers and just added in the one (mandatory) marker needed since the optional markers weren't correct any more.