RULEAPI-822 Don't consider sonar-css that is deprecated, when computing coverage
sonar-css-plugin is no longer used and the sonar-css repository is archived. Rules that are only in
sonar-css should be considered as deleted and rules that are moved in sonarJS should be displayed
as in sonarJS.
The current coverage format cannot represent perfectly this repository change, so removing sonar-css
from the list seems like the simplest approach to get a good enough result.